requirements.txt
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2
torch>=2.0
PyYAML>=6.0
matplotlib>=3.7
openpyxl>=3.1
joblib>=1.2
pytest>=7.0
